requirements.txt
numpy==1.24.1
onnx==1.14.1
Pillow==9.3.0
Pillow==10.1.0
scikit_learn==1.3.2
torch==2.2.0.dev20231008
torchvision==0.17.0.dev20231008
visdom==0.2.4
